Update README.md
# Archipulse Frontend Website

Explore the world of real estate with [Your Real Estate Firm Name]'s frontend website. This captivating online platform features a thoughtfully designed introduction, a comprehensive overview of our services in the "What We Offer" section, a showcase of available properties, and a convenient footer for quick access to contact details and more. The website is built with a focus on user experience, utilizing modern frontend technologies to ensure a seamless and visually appealing journey for every visitor.